Filters are now proper components
